refactor: extract Devtool commands
AI-Generated: ChatGPT (OpenAI)
