+++ client/src/__tests__/unit-tests/ui/devtool-commands.test.ts
@@ -9,7 +9,7 @@ import { BitbakeWorkspace } from '../../../ui/BitbakeWorkspace'
 import { BitBakeProjectScanner } from '../../../driver/BitBakeProjectScanner'
 import { mockVscodeExtensionContext } from '../../utils/vscodeMock'
 import { BitbakeDriver } from '../../../driver/BitbakeDriver'
-import { addDevtoolDebugBuild, registerDevtoolCommands } from '../../../ui/BitbakeCommands'
+import { addDevtoolDebugBuild, registerDevtoolCommands } from '../../../ui/DevtoolCommands'
 import { clientNotificationManager } from '../../../ui/ClientNotificationManager'
 import * as BitbakeTerminal from '../../../ui/BitbakeTerminal'
 import * as ProcessUtils from '../../../utils/ProcessUtils'
